+++ rust/src/libponyc/codegen/gencontrol.rs
@@ -1651,8 +1651,8 @@ pub unsafe extern "C" fn gen_if(mut c: *mut compile_t, mut ast: *mut ast_t) -> L
         [&mut cond, &mut left, &mut right, 0 as *mut *mut ast_t];
     ast_get_children(
         ast,
-        (::core::mem::size_of::<[*mut *mut ast_t; 4]>() as libc::c_ulong)
-            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
+        ::core::mem::size_of::<[*mut *mut ast_t; 4]>()
+            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>())
             .wrapping_sub(1),
         children.as_mut_ptr(),
     );
@@ -1780,7 +1780,7 @@ pub unsafe extern "C" fn gen_iftype(mut c: *mut compile_t, mut ast: *mut ast_t) 
     ast_get_children(
         ast,
         (::core::mem::size_of::<[*mut *mut ast_t; 3]>() as libc::c_ulong)
-            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
+            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>())
             .wrapping_sub(1),
         children.as_mut_ptr(),
     );
@@ -1795,8 +1795,8 @@ pub unsafe extern "C" fn gen_iftype(mut c: *mut compile_t, mut ast: *mut ast_t) 
     ];
     ast_get_children(
         left,
-        (::core::mem::size_of::<[*mut *mut ast_t; 4]>() as libc::c_ulong)
-            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
+        ::core::mem::size_of::<[*mut *mut ast_t; 4]>()
+            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>())
             .wrapping_sub(1),
         children_0.as_mut_ptr(),
     );
@@ -1822,8 +1822,8 @@ pub unsafe extern "C" fn gen_while(mut c: *mut compile_t, mut ast: *mut ast_t) -
         [&mut cond, &mut body, &mut else_clause, 0 as *mut *mut ast_t];
     ast_get_children(
         ast,
-        (::core::mem::size_of::<[*mut *mut ast_t; 4]>() as libc::c_ulong)
-            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
+        ::core::mem::size_of::<[*mut *mut ast_t; 4]>()
+            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>())
             .wrapping_sub(1),
         children.as_mut_ptr(),
     );
@@ -1938,8 +1938,8 @@ pub unsafe extern "C" fn gen_repeat(mut c: *mut compile_t, mut ast: *mut ast_t) 
         [&mut body, &mut cond, &mut else_clause, 0 as *mut *mut ast_t];
     ast_get_children(
         ast,
-        (::core::mem::size_of::<[*mut *mut ast_t; 4]>() as libc::c_ulong)
-            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
+        ::core::mem::size_of::<[*mut *mut ast_t; 4]>()
+            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>())
             .wrapping_sub(1),
         children.as_mut_ptr(),
     );
@@ -2187,8 +2187,8 @@ pub unsafe extern "C" fn gen_try(mut c: *mut compile_t, mut ast: *mut ast_t) -> 
     ];
     ast_get_children(
         ast,
-        (::core::mem::size_of::<[*mut *mut ast_t; 4]>() as libc::c_ulong)
-            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
+        ::core::mem::size_of::<[*mut *mut ast_t; 4]>()
+            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>())
             .wrapping_sub(1),
         children.as_mut_ptr(),
     );
@@ -2302,7 +2302,7 @@ pub unsafe extern "C" fn gen_disposing_block_can_error(
     ast_get_children(
         ast,
         (::core::mem::size_of::<[*mut *mut ast_t; 3]>() as libc::c_ulong)
-            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
+            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>())
             .wrapping_sub(1),
         children.as_mut_ptr(),
     );
@@ -2402,7 +2402,7 @@ pub unsafe extern "C" fn gen_disposing_block_cant_error(
     ast_get_children(
         ast,
         (::core::mem::size_of::<[*mut *mut ast_t; 3]>() as libc::c_ulong)
-            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
+            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>())
             .wrapping_sub(1),
         children.as_mut_ptr(),
     );
@@ -2515,7 +2515,7 @@ pub unsafe extern "C" fn attach_branchweights_metadata(
 ) {
     let mut alloc_index: usize = ponyint_pool_index(
         (count.wrapping_add(1 as libc::c_int as libc::c_uint) as libc::c_ulong)
-            .wrapping_mul(::core::mem::size_of::<LLVMValueRef>() as libc::c_ulong),
+            .wrapping_mul(::core::mem::size_of::<LLVMValueRef>()),
     );
     let mut params: *mut LLVMValueRef = ponyint_pool_alloc(alloc_index) as *mut LLVMValueRef;
     let str: [libc::c_char; 15] =
