+++ rust/src/libponyc/codegen/genmatch.rs
@@ -1820,7 +1820,7 @@ unsafe extern "C" fn eq_param_type(mut c: *mut compile_t, mut pattern: *mut ast_
     ast_get_children(
         (*fun).ast,
         (::core::mem::size_of::<[*mut *mut ast_t; 7]>() as libc::c_ulong)
-            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
+            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>())
             .wrapping_sub(1),
         children.as_mut_ptr(),
     );
@@ -2615,8 +2615,8 @@ pub unsafe extern "C" fn gen_match(mut c: *mut compile_t, mut ast: *mut ast_t) -
     ];
     ast_get_children(
         ast,
-        (::core::mem::size_of::<[*mut *mut ast_t; 4]>() as libc::c_ulong)
-            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
+        ::core::mem::size_of::<[*mut *mut ast_t; 4]>()
+            .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>())
             .wrapping_sub(1),
         children.as_mut_ptr(),
     );
@@ -2674,8 +2674,8 @@ pub unsafe extern "C" fn gen_match(mut c: *mut compile_t, mut ast: *mut ast_t) -
             [&mut pattern, &mut guard, &mut body, 0 as *mut *mut ast_t];
         ast_get_children(
             the_case,
-            (::core::mem::size_of::<[*mut *mut ast_t; 4]>() as libc::c_ulong)
-                .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
+            ::core::mem::size_of::<[*mut *mut ast_t; 4]>()
+                .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>())
                 .wrapping_sub(1),
             children_0.as_mut_ptr(),
         );
